+++ rstar/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+## Added
+- Added method `RTree::drain()`.
+
 ## Changed
 - fixed all clippy lint issues
 - Fixed error when setting MIN_SIZE = 1 in `RTreeParams` and added assert for positive MIN_SIZE