+++ rstar/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 - Added method `RTree::drain()`.
+- Changed license field to [SPDX 2.1 license expression](https://spdx.dev/spdx-specification-21-web-version/#h.jxpfx0ykyb60)
 
 ## Changed
 - fixed all clippy lint issues