+++ rstar/CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Changed
 - fixed all clippy lint issues
 - Fixed error when setting MIN_SIZE = 1 in `RTreeParams` and added assert for positive MIN_SIZE
+- Removed the `Copy` bound from `Point` and `Envelope`.
 
 # 0.9.3
 ## Changed