+++ rstar/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Added
 - Add `CachedEnvelope` combinator which simplifies memoizing envelope computations. ([PR](https://github.com/georust/rstar/pull/118))
 
+## Changed
+- Increase our MSRV to Rust 1.63 following that of the `geo` crate.  ([PR](https://github.com/georust/rstar/pull/124))
+
 # 0.10.0
 
 ## Added