+++ rstar/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Added
 - Add `CachedEnvelope` combinator which simplifies memoizing envelope computations. ([PR](https://github.com/georust/rstar/pull/118))
+- `Point` is now implemented as const generic for any length of `RTreeNum` array
 
 ## Changed
 - Increase our MSRV to Rust 1.63 following that of the `geo` crate.  ([PR](https://github.com/georust/rstar/pull/124))
@@ -11,7 +12,6 @@
 ## Added
 - Added method `RTree::drain()`.
 - Changed license field to [SPDX 2.1 license expression](https://spdx.dev/spdx-specification-21-web-version/#h.jxpfx0ykyb60)
-- `Point` is now implemented as const generic for any length of `RTreeNum` array
 
 ## Changed
 - fixed all clippy lint issues